+++ changelog.md
@@ -1,4 +1,8 @@
 # bpd cUI Light changelog
+# [0.1.13] WIP
+# [0.1.12] 2020-08-12
+* [Fixed] Mutation handler does not update component
+* [Added] Option 'height' to switch component
 
 # [0.1.11] 2020-08-10
 * [Fixed] Adjusted all component to use new API and events system
@@ -16,8 +20,10 @@
 
 # [0.1.8] 2020-08-06
 * [Fixed] Offcanvas background
+
 # [0.1.7] 2020-08-06
 * [Fixed] Dialog and offcanvas background when opened
+
 # [0.1.6] 2020-08-06
 * [Fixed] Offset element now supports root element
 * [Fixed] Blinking dialog and offcanvas when closed via keys or outside click